requirements.txt
paho-mqtt==2.0.0
psutil==5.9.8
pyyaml==6.0.1
schedule==1.2.1